requirements.txt
pydicom 
monai 
einops 
timm
pytorch